refactor(intents): address PR #249 review feedback (#235)
- backup-site: reclassify .modifiesContent → .createsContent. BackupCommand
  is git add -A → commit → push to a draft branch; it snapshots the working
  tree outward, it does not alter existing site source. Tighten the
  OperationSideEffect docs so the taxonomy is about write reach, not "source".
- mcpToolName: replace the self-defeating "all nil" snapshot test with a
  positive settability test (the nil snapshot blocked ever filling the field).
- FakeContentOps: extract from ContentIntentsTests to Support/ and share with
  the behavioral suite (removes the RoutingContentOps duplicate); add the same
  @unchecked Sendable thread-safety doc FakeOperations carries.
- readsDoNotMutate: also inject FakeOperations under SiteOperationsOverride and
  assert deploy/backup/audit stay empty — a read intent wrongly calling a site
  op would now fail, not just a content-create.
- edit-content: add TODO(#239/#250) checkpoint comments at the descriptor and
  the test value table so the requiresConfirmation flip isn't missed.
- Remove the implementation plan doc (agent scaffolding with hardcoded local
  paths); the spec doc remains as the design artifact.

172 intents tests pass.
