+++ Tests/AnglesiteIntentsTests/OperationDescriptorTests.swift
@@ -44,7 +44,7 @@ extension AppIntentsTests {
             }
             let expected: [String: Expected] = [
                 "deploy-site": .init(sideEffect: .publishes, requiresConfirmation: true, isCancellable: true, resultShape: .entity("SiteEntity")),
-                "backup-site": .init(sideEffect: .modifiesContent, requiresConfirmation: false, isCancellable: true, resultShape: .entity("SiteEntity")),
+                "backup-site": .init(sideEffect: .createsContent, requiresConfirmation: false, isCancellable: true, resultShape: .entity("SiteEntity")),
                 "audit-site": .init(sideEffect: .readOnly, requiresConfirmation: false, isCancellable: true, resultShape: .entity("SiteEntity")),
                 "open-site": .init(sideEffect: .readOnly, requiresConfirmation: false, isCancellable: false, resultShape: .none),
                 "search-content": .init(sideEffect: .readOnly, requiresConfirmation: false, isCancellable: false, resultShape: .entities("ContentMatchEntity")),
@@ -52,6 +52,8 @@ extension AppIntentsTests {
                 "preview-site": .init(sideEffect: .readOnly, requiresConfirmation: false, isCancellable: false, resultShape: .none),
                 "add-page": .init(sideEffect: .createsContent, requiresConfirmation: false, isCancellable: true, resultShape: .entity("PageEntity")),
                 "add-post": .init(sideEffect: .createsContent, requiresConfirmation: false, isCancellable: true, resultShape: .entity("PostEntity")),
+                // TODO(#239/#250): flip requiresConfirmation to true when the EditContentIntent
+                // confirmation gate lands — this assertion passes on stale data until then.
                 "edit-content": .init(sideEffect: .modifiesContent, requiresConfirmation: false, isCancellable: true, resultShape: .none),
             ]
             #expect(expected.count == AnglesiteOperations.all.count)
@@ -64,9 +66,14 @@ extension AppIntentsTests {
             }
         }
 
-        @Test("mcpToolName is nil for all current entries (forward-looking field)")
-        func mcpToolNamesUnset() {
-            #expect(AnglesiteOperations.all.allSatisfy { $0.mcpToolName == nil })
+        @Test("mcpToolName can be set on a descriptor (forward-looking field)")
+        func mcpToolNameIsSettable() {
+            let d = OperationDescriptor(
+                operationID: "x", displayName: "X", intentTypeName: "XIntent",
+                sideEffect: .readOnly, requiresConfirmation: false,
+                isCancellable: false, resultShape: .none, mcpToolName: "x_tool"
+            )
+            #expect(d.mcpToolName == "x_tool")
         }
     }
 }
